Added basic functionality to have a dedicated validation endpoint which do only validation.
Summary:
For clients it should be an endpoint available where the client can validate an app or pod. Additionally to the regular endpoints
there is the functionality of a strict validation. This means, that you would receive an error message if you provide json properties, which
are unknown in the raml definition.

Test Plan: sbt test

Reviewers: aquamatthias, meichstedt, jasongilanfarr, zen-dog

Subscribers: marathon-dev, marathon-team

JIRA Issues: MARATHON-7037

Differential Revision: https://phabricator.mesosphere.com/D698